requirements.txt
fastapi
uvicorn
torch
torchvision
--extra-index-url https://download.pytorch.org/whl/cpu
transformers
pillow
python-multipart
numpy